+++ packages/textlint/src/engine/textlint-module-loader.ts
@@ -12,11 +12,11 @@ import { TextLintModuleResolver } from "./textlint-module-resolver";
 import { TextLintModuleMapper } from "./textlint-module-mapper";
 import { Config } from "../config/config";
 import {
-    normalizeFilterRuleKey,
-    normalizePluginKey,
-    normalizeRuleKey,
-    normalizeRulePresetKey
-} from "../config/config-key-normalizer";
+    normalizeTextlintFilterRuleKey,
+    normalizeTextlintPluginKey,
+    normalizeTextlintRuleKey,
+    normalizeTextlintRulePresetKey
+} from "@textlint/types";
 
 export class TextLintModuleLoader extends EventEmitter {
     moduleResolver: TextLintModuleResolver;
@@ -92,7 +92,7 @@ export class TextLintModuleLoader extends EventEmitter {
         const pkgPath = this.moduleResolver.resolvePluginPackageName(pluginName);
         debug("Loading rules from plugin: %s", pkgPath);
         const plugin = interopRequire(pkgPath);
-        const pluginNameWithoutPrefix = normalizePluginKey(pluginName);
+        const pluginNameWithoutPrefix = normalizeTextlintPluginKey(pluginName);
         // Notes: plugins not support "rules" and "rulesConfig"
         // https://github.com/textlint/textlint/issues/291
         if (plugin.hasOwnProperty("rules")) {
@@ -123,7 +123,7 @@ For more details. See https://github.com/textlint/textlint/blob/master/docs/plug
         It mean that "ruleA" is defined as "preset-gizmo/ruleA"
 
          */
-        const presetRuleNameWithoutPrefix = normalizeRulePresetKey(presetName);
+        const presetRuleNameWithoutPrefix = normalizeTextlintRulePresetKey(presetName);
         // ignore plugin's rule
         if (isPluginRuleKey(presetRuleNameWithoutPrefix)) {
             Logger.warn(`${presetRuleNameWithoutPrefix} is Plugin's rule. This is unknown case, please report issue.`);
@@ -162,7 +162,7 @@ For more details. See https://github.com/textlint/textlint/blob/master/docs/plug
         }
         // ignore already defined rule
         // ignore rules from rulePaths because avoid ReferenceError is that try to require.
-        const definedRuleName = normalizeRuleKey(ruleName);
+        const definedRuleName = normalizeTextlintRuleKey(ruleName);
         // ignore plugin's rule
         if (isPluginRuleKey(definedRuleName)) {
             Logger.warn(`${definedRuleName} is Plugin's rule. This is unknown case, please report issue.`);
@@ -197,7 +197,7 @@ For more details. See https://github.com/textlint/textlint/blob/master/docs/plug
             this.emit(TextLintModuleLoader.Event.filterRule, ruleEntry);
             return;
         }
-        const definedRuleName = normalizeFilterRuleKey(ruleName);
+        const definedRuleName = normalizeTextlintFilterRuleKey(ruleName);
         // ignore plugin's rule
         if (isPluginRuleKey(definedRuleName)) {
             Logger.warn(`${definedRuleName} is Plugin's rule. This is unknown case, please report issue.`);